nl2sql_standalone_AzureAI/requirements.txt
# Azure AI Agent Service dependencies
azure-ai-projects>=1.0.0
azure-identity>=1.19.0

# Database connectivity
pyodbc>=5.2.0

# Utilities
python-dotenv>=1.0.0
requests>=2.32.0

# Optional: for Streamlit UI
streamlit>=1.40.0
pandas>=2.2.0
